requirements.txt
requests
flask
pycryptodome